Refactor: 패키지 명 변경 및 패키지 생성
org.example에서 org.chess로 패키지 이름 변경, org.chess 하위에 pieses 패키지 생성하고 기존의 Pawn 클래스를 pieses 패키지 하위로 이동
