+++ app/src/main/java/org/chess/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.chess.pieces;
 
 public class Pawn {
